feat!: Deprecate Search Guard variant
This integration has been unmaintained for over 2 years. It is still
based on the deprecated "Classic" Search Guard[1] and does not support
the version 8 of the Elastic stack. Migrating to the newer Search Guard
FLX would require a significant overhaul which doesn't seem justified
considered the lack of traction for Search Guard inside this project.

[1]: #826