fix(setup): Set curl --resolve flag conditionally
curl 8.x no longer treats CURLOPT_RESOLVE parse failures as a warning.
Because 'elasticsearch' is not an IP address, using it as the ADDR
segment in the --resolve flag is illegal.

In 7.x, curl would print a warning but fall back to the hostname in the
URL, as if --resolve had not been set.

This issue went under the radar because:

1. curl 8.x is used in Elastic v8 images (Ubuntu base[1]), whereas our
   default Elastic v9 images still ship with curl 7.x (RHEL base[2]).
2. The warning was not printed due to the usage of the -s flag.

Fixes #1137

[1]: https://github.com/elastic/elasticsearch/blob/v8.19.8/build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/DockerBase.java#L16
[2]: https://github.com/elastic/elasticsearch/blob/v9.2.2/build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/DockerBase.java#L17